chore: move vpp related code into dedicated files
