+++ server/service/software_installers.go
@@ -400,65 +400,3 @@ func (svc *Service) HasSelfServiceSoftwareInstallers(ctx context.Context, host *
 
 	return svc.ds.HasSelfServiceSoftwareInstallers(ctx, host.Platform, host.TeamID)
 }
-
-//////////////////////////////////////////////////////////////////////////////
-// Get App Store Software
-//////////////////////////////////////////////////////////////////////////////
-
-type getAppStoreSoftwareRequest struct {
-	TeamID uint `query:"team_id"`
-}
-
-type getAppStoreSoftwareResponse struct {
-	AppStoreApps []*fleet.VPPApp `json:"app_store_apps"`
-	Err          error           `json:"error,omitempty"`
-}
-
-func (r getAppStoreSoftwareResponse) error() error { return r.Err }
-
-func getAppStoreSoftwareEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (errorer, error) {
-	req := request.(*getAppStoreSoftwareRequest)
-	apps, err := svc.GetAppStoreSoftware(ctx, &req.TeamID)
-	if err != nil {
-		return &getAppStoreSoftwareResponse{Err: err}, nil
-	}
-
-	return &getAppStoreSoftwareResponse{AppStoreApps: apps}, nil
-}
-
-func (svc *Service) GetAppStoreSoftware(ctx context.Context, teamID *uint) ([]*fleet.VPPApp, error) {
-	// skipauth: No authorization check needed due to implementation returning
-	// only license error.
-	svc.authz.SkipAuthorization(ctx)
-
-	return nil, fleet.ErrMissingLicense
-}
-
-type addAppStoreAppRequest struct {
-	TeamID     *uint  `json:"team_id"`
-	AppStoreID string `json:"app_store_id"`
-}
-
-type addAppStoreAppResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r addAppStoreAppResponse) error() error { return r.Err }
-
-func addAppStoreAppEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (errorer, error) {
-	req := request.(*addAppStoreAppRequest)
-	err := svc.AddAppStoreApp(ctx, req.TeamID, req.AppStoreID)
-	if err != nil {
-		return &addAppStoreAppResponse{Err: err}, nil
-	}
-
-	return &addAppStoreAppResponse{}, nil
-}
-
-func (svc *Service) AddAppStoreApp(ctx context.Context, teamID *uint, adamID string) error {
-	// skipauth: No authorization check needed due to implementation returning
-	// only license error.
-	svc.authz.SkipAuthorization(ctx)
-
-	return fleet.ErrMissingLicense
-}
